[DEL] page file deleted. 'Page' class moved to comic.py file.
