+++ pynocchio/main_window_model.py
@@ -76,7 +76,7 @@ class MainWindowModel(QtCore.QObject):
         try:
             loader.load(filename)
         except NoDataFindException as exc:
-            from pynocchio.page import Page
+            from pynocchio.comic import Page
             logger.exception('Error in load comic! %s' % exc)
             q_file = QtCore.QFile(":/icons/notCover.png")
             q_file.open(QtCore.QIODevice.ReadOnly)
