+++ test/test_comic_page_handler.py
@@ -17,7 +17,7 @@
 
 from unittest import TestCase
 
-from pynocchio.comic import Comic
+from pynocchio.comic import Comic, Page
 from pynocchio.page import Page
 from pynocchio.comic_page_handler import ComicPageHandler
